Centralize asset url parsing.
