+++ Assets/Source/Application/Services/AssetUrlHelper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: afd27af5efe91bb48a11886583ddd46a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
